+++ src/Typeahead.svelte
@@ -1,8 +1,10 @@
 <script>
   /**
-   * @typedef {any} Item
-   * @event {{ selectedIndex: number; selected: Item; }} select
-   * @slot {{ result: { index: number; original: Item; score: number; string: string; } }}
+   * @typedef {string | number | Record<string, any>} Item
+   * @typedef {{ original: Item; index: number; score: number; string: string; }} FuzzyResult
+   * @event {{ searched: string; selected: Item; selectedIndex: number; original: Item; originalIndex: number; }} select
+   * @event {any} clear
+   * @slot {{ result: FuzzyResult; index: number }}
    */
 
   export let id = "typeahead-" + Math.random().toString(36);
@@ -13,7 +15,27 @@
 
   /** @type {(item: Item) => Item} */
   export let extract = (item) => item;
+
+  /** @type {(item: Item) => Item} */
+  export let disable = (item) => false;
+
+  /** @type {(item: Item) => Item} */
+  export let filter = (item) => false;
+
+  /** Set to `false` to prevent the first result from being selected */
   export let autoselect = true;
+
+  /**
+   * Set to `keep` to keep the search field unchanged after select, set to `clear` to auto-clear search field
+   * @type {"update" | "clear" | "keep"}
+   */
+  export let inputAfterSelect = "update";
+
+  /** @type {FuzzyResult[]} */
+  export let results = [];
+
+  /** Set to `true` to re-focus the input after selecting a result */
+  export let focusAfterSelect = false;
 
   import fuzzy from "fuzzy";
   import Search from "svelte-search";
@@ -21,8 +43,8 @@
 
   const dispatch = createEventDispatcher();
 
-  let comboboxRef = undefined;
-  let searchRef = undefined;
+  let comboboxRef = null;
+  let searchRef = null;
   let hideDropdown = false;
   let selectedIndex = -1;
   let prevResults = "";
@@ -32,120 +54,112 @@
       selectedIndex = 0;
     }
 
+    if (prevResults !== resultsId) {
+      hideDropdown = results.length === 0;
+    }
+
     prevResults = resultsId;
   });
 
   async function select() {
-    value = extract(results[selectedIndex].original);
-    dispatch("select", { selectedIndex, selected: value });
+    const result = results[selectedIndex];
+    const selectedValue = extract(result.original);
+    const searchedValue = value;
+
+    if (inputAfterSelect == "clear") value = "";
+    if (inputAfterSelect == "update") value = selectedValue;
+
+    dispatch("select", {
+      selectedIndex,
+      searched: searchedValue,
+      selected: selectedValue,
+      original: result.original,
+      originalIndex: result.index,
+    });
+
     await tick();
-    searchRef.focus();
+
+    if (focusAfterSelect) searchRef.focus();
+
     hideDropdown = true;
   }
 
   $: options = { pre: "<mark>", post: "</mark>", extract };
   $: results = fuzzy
     .filter(value, data, options)
-    .filter(({ score }) => score > 0);
+    .filter(({ score }) => score > 0)
+    .filter((result) => !filter(result.original))
+    .map((result) => ({ ...result, disabled: disable(result.original) }));
   $: resultsId = results.map((result) => extract(result.original)).join("");
 </script>
 
-<style>
-  .display-block {
-    display: block;
-  }
-  
-  .svelte-typeahead {
-    position: relative;
-  }
-
-  :global(.svelte-typeahead.dropdown .svelte-search input) {
-    border-bottom-right-radius: 0;
-    border-bottom-left-radius: 0;
-  }
-
-  :global(.svelte-search input) {
-    border: 0;
-    background: none;
-    width: 100%;
-    font: inherit;
-    font-size: 1.5rem;
-    padding: 1rem;
-    border: 2px solid #e0e0e0;
-    border-radius: 0.25rem;
-  }
-
-  :global(.svelte-search input:focus) {
-    outline: 0;
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-    border-color: #0f62fe;
-  }
-</style>
-
 <svelte:window
-  on:click="{({ target }) => {
+  on:click={({ target }) => {
     if (!hideDropdown && results.length > 0 && comboboxRef && !comboboxRef.contains(target)) {
       hideDropdown = true;
     }
-  }}"
+  }}
 />
 
 <div
-  bind:this="{comboboxRef}"
+  data-svelte-typeahead
+  bind:this={comboboxRef}
   role="combobox"
   aria-haspopup="listbox"
   aria-owns="{id}-listbox"
   class="dropdown"
-  
   aria-expanded="{!hideDropdown && results.length > 0}"
   id="{id}"
 >
   <Search
     {...$$restProps}
-    bind:this="{searchRef}"
+    bind:ref={searchRef}
     aria-autocomplete="list"
     aria-controls="{id}-listbox"
     aria-labelledby="{id}-label"
     aria-activedescendant=""
-    id="{id}"
     bind:value
+    on:type
     on:input
     on:change
     on:focus
-    on:focus="{() => {
+    on:focus={() => {
       hideDropdown = false;
-    }}"
-    on:clear="{() => {
+    }}
+    on:clear
+    on:clear={() => {
       hideDropdown = false;
-    }}"
+    }}
     on:blur
     on:keydown
-    on:keydown="{({ key }) => {
-      switch (key) {
+    on:keydown={(e) => {
+      switch (e.key) {
         case 'Enter':
           select();
           break;
         case 'ArrowDown':
+          e.preventDefault();
           selectedIndex += 1;
           if (selectedIndex === results.length) {
             selectedIndex = 0;
           }
           break;
         case 'ArrowUp':
+          e.preventDefault();
           selectedIndex -= 1;
           if (selectedIndex < 0) {
             selectedIndex = results.length - 1;
           }
           break;
         case 'Escape':
+          e.preventDefault();
           value = '';
           searchRef.focus();
           hideDropdown = true;
           break;
       }
-    }}"
+    }}
   />
-
   <div class="dropdown-menu" class:display-block="{!hideDropdown}" id="{id}-listbox" role="menu">
     {#if !hideDropdown && results.length > 0}
         <div class="dropdown-content">
@@ -153,10 +167,13 @@
             <a id="{id}-result"
                class="dropdown-item"
                class:is-active="{selectedIndex === i}"
+               class:disabled={result.disabled}
                aria-selected="{selectedIndex === i}"
                on:click="{() => {
-                 selectedIndex = i;
-                 select();
+                 if (!result.disabled) {
+                   selectedIndex = i;
+                   select();
+                 }
                }}"
               >
                 {@html result.string}
@@ -166,3 +183,77 @@
     {/if}
   </div>
 </div>
+
+<style>
+  [data-svelte-typeahead] {
+    position: relative;
+    background-color: #fff;
+  }
+
+  ul {
+    position: absolute;
+    top: 100%;
+    left: 0;
+    width: 100%;
+    padding: 0;
+    list-style: none;
+    background-color: inherit;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+  }
+
+  li {
+    padding: 0.25rem 1rem;
+    cursor: pointer;
+  }
+
+  li:not(:last-of-type) {
+    border-bottom: 1px solid #e0e0e0;
+  }
+
+  li:hover {
+    background-color: #e5e5e5;
+  }
+
+  .selected {
+    background-color: #e5e5e5;
+  }
+
+  .selected:hover {
+    background-color: #cacaca;
+  }
+
+  .disabled {
+    opacity: 0.4;
+    cursor: not-allowed;
+  }
+
+  :global([data-svelte-search] label) {
+    margin-bottom: 0.25rem;
+    display: inline-flex;
+    font-size: 0.875rem;
+  }
+
+  :global([data-svelte-search] input) {
+    width: 100%;
+    padding: 0.5rem 0.75rem;
+    background: none;
+    font-size: 1rem;
+    border: 0;
+    border-radius: 0;
+    border: 1px solid #e5e5e5;
+  }
+
+  :global([data-svelte-search] input:focus) {
+    outline-color: #0f62fe;
+    outline-offset: 2px;
+    outline-width: 1px;
+  }
+
+  .display-block {
+    display: block;
+  }
+  
+  .svelte-typeahead {
+    position: relative;
+  }
+</style>